requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncmy==0.2.9
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi_nextauth_jwt==2.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jose==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
passlib==1.7.4
pillow==10.4.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
qrcode==7.4.2
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
types-passlib==1.7.7.20240327
types-pyasn1==0.6.0.20240402
types-python-jose==3.3.4.20240106
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
